+++ _benchmarks/logrus_test.go
@@ -23,5 +23,4 @@ func printLogrus(i int) {
 	logrus.Errorf("[%d] This is an error message", i)
 	logrus.Warnf("[%d] This is a warning message", i)
 	logrus.Infof("[%d] This is an info message", i)
-	logrus.Debugf("[%d] This is a debug message", i)
 }