+++ _examples/customize-output/main.go
@@ -26,7 +26,7 @@ func main() {
 	*/
 	golog.Debugf("This is a %s with data (debug prints the stacktrace too)", "message", golog.Fields{
 		"username": "kataras",
-	})
+	}) // If more than one golog.Fields passed, then they are merged into a single map.
 
 	/* Example Output:
 	{
@@ -36,12 +36,23 @@ func main() {
 	    "fields": {
 	        "home": "https://iris-go.com"
 	    }
+		"stacktrace": [...]
 	}
 	*/
 	golog.Infof("An info message", golog.Fields{"home": "https://iris-go.com"})
 
 	golog.Warnf("Hey, warning here")
 	golog.Errorf("Something went wrong!")
+
+	// You can also pass custom structs, like normally you would do.
+	type myCustomData struct {
+		Username string `json:"username"`
+		Email    string `json:"email"`
+	}
+	golog.Fatalf("A fatal error for %s screen!", "home", golog.Fields{"data": myCustomData{
+		Username: "kataras",
+		Email:    "kataras2006@hotmail.com",
+	}})
 }
 
 /* Manually, use it for any custom format: